OGM-1002 Move Neo4j tests
