+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/neo4j/Neo4jJtaModuleMemberRegistrationIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.jboss;
+package org.hibernate.ogm.test.integration.neo4j;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -15,8 +15,9 @@ import java.io.InputStream;
 import org.hibernate.ogm.datastore.neo4j.Neo4j;
 import org.hibernate.ogm.datastore.neo4j.Neo4jProperties;
 import org.hibernate.ogm.jpa.HibernateOgmPersistence;
-import org.hibernate.ogm.test.integration.jboss.model.Member;
-import org.hibernate.ogm.test.integration.jboss.util.ModuleMemberRegistrationDeployment;
+import org.hibernate.ogm.test.integration.testcase.ModuleMemberRegistrationScenario;
+import org.hibernate.ogm.test.integration.testcase.model.Member;
+import org.hibernate.ogm.test.integration.testcase.util.ModuleMemberRegistrationDeployment;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
